+++ ricaun.Revit.DB.Shape.Tests/Utils/ElementIdUtils.cs
@@ -2,6 +2,9 @@
 
 namespace ricaun.Revit.DB.Shape.Tests.Utils
 {
+    /// <summary>
+    /// Use this ElementId only outside a document context.
+    /// </summary>
     public static class ElementIdUtils
     {
         public static ElementId MaterialId => new ElementId(BuiltInCategory.OST_Materials);