+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] / 2024-09-23
+### Shapes
+- Add `FaceExtension` convert face to solid.
+
 ## [0.3.2] / 2024-06-14
 ### Shapes
 - Update Colors using the `System.Windows.Media.Colors` as reference.
@@ -74,6 +78,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Create Project `ricaun.Revit.DB`
 
 [vNext]: ../../compare/0.1.0...HEAD
+[0.4.0]: ../../compare/0.3.2...0.4.0
 [0.3.2]: ../../compare/0.3.1...0.3.2
 [0.3.1]: ../../compare/0.3.0...0.3.1
 [0.3.0]: ../../compare/0.2.0...0.3.0